ENH: Be able to flatten hierarchical JSON
This is helpful when more than the current BIDS set of information is
available in the JSON structs (e.g. a DICOM dump). Nested "dicts" will
appear as a flat list of columns with the key hierarchy presented as a
concatenated string joined by ":'. For example:

time -> samples -> TriggerTime

yields

Parameter Value[time:samples:TriggerTime]

This change should have no effect on a plain BIDS metadata.